stress-bpf: clean up 32 bit pointer to uint64_t cast warning
Cast 32 bit pointers to uintptr_t before casting to uint64_t to
silence pointer to int cast warnings:

stress-bpf.c: In function ‘stress_bpf_prog_load’:
stress-bpf.c:185:26: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
  185 |                 .insns = (uint64_t)insns,
      |                          ^
stress-bpf.c:187:28: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
  187 |                 .license = (uint64_t)"GPL",

+++ stress-bpf.c
@@ -182,9 +182,9 @@ static inline int ALWAYS_INLINE stress_bpf_prog_load(
 {
 	union bpf_attr attr = {
 		.prog_type = type,
-		.insns = (uint64_t)insns,
+		.insns = (uint64_t)(uintptr_t)insns,
 		.insn_cnt = insn_cnt,
-		.license = (uint64_t)"GPL",
+		.license = (uint64_t)(uintptr_t)"GPL",
 		.log_buf = 0,
 		.log_size = 0,
 		.log_level = 0,
